engine: extract dispatch::hunters (#159)
